+++ src/main/resources/application.properties
@@ -1,0 +1,19 @@
+spring.datasource.url=
+spring.datasource.username=
+spring.datasource.password=
+spring.security.oauth2.resourceserver.jwt.issuer-uri=
+springdoc.swagger-ui.operationsSorter=
+elasticsearch.hostname=
+elasticsearch.port=
+elasticsearch.index-name=
+kafka.consumer.host=
+kafka.consumer.group=
+kafka.consumer.topic=
+kafka.publisher.host=
+springdoc.show-actuator=
+spring.profiles.active=
+endpoint.token-endpoint=
+endpoint.handle-endpoint=
+token.secret=
+token.id=
+token.grant-type=
